+++ src-tauri/src/lib.rs
@@ -22,11 +22,7 @@ use app::AppState;
 
 use std::time::Duration;
 // use tauri::api::path::cache_dir;
-use tauri::{
-    menu::{MenuBuilder, MenuItemBuilder},
-    tray::TrayIconBuilder,
-    Manager,
-};
+use tauri::{menu::MenuItemBuilder, Manager};
 use tauri_plugin_autostart::MacosLauncher;
 // use tauri_plugin_log::LogTarget;
 
@@ -102,7 +98,7 @@ fn build_and_run_app(app: AppState) {
 
             let menu = tauri::menu::MenuBuilder::new(app).items(&[&quit]).build()?;
 
-            let tray = tauri::tray::TrayIconBuilder::with_id("tray")
+            let _tray = tauri::tray::TrayIconBuilder::with_id("tray")
                 .tooltip("Pachtop")
                 .icon(icon)
                 .menu(&menu)
@@ -113,7 +109,7 @@ fn build_and_run_app(app: AppState) {
                     _ => {}
                 })
                 .on_tray_icon_event(|tray, event| match event {
-                    tauri::tray::TrayIconEvent::Click { id, button, .. } => match button {
+                    tauri::tray::TrayIconEvent::Click { button, .. } => match button {
                         tauri::tray::MouseButton::Left => {
                             let app = tray.app_handle();
                             if let Some(window) = app.get_webview_window("main") {